+++ src/utils/turndown-rules/images-rule.ts
@@ -4,6 +4,7 @@ import { filterByNodeName } from './filter-by-nodename';
 import { getAttributeProxy } from './get-attribute-proxy';
 import { OutputFormat } from './../../output-format';
 import { isHeptaOrObsidianOutput } from './../is-hepta-or-obsidian-output';
+import { ImageSizeFormat } from './../../image-size-format';
 
 export const imagesRule = {
   filter: filterByNodeName('IMG'),
@@ -22,21 +23,22 @@ export const imagesRule = {
     } else if (yarleOptions.urlEncodeFileNamesAndLinks) {
       realValue = encodeURI(realValue);
     }
-    let sizeString = (widthParam || heightParam) ? ` =${widthParam}x${heightParam}` : '';
+    if (yarleOptions.keepImageSize){
+      let sizeString = (widthParam || heightParam) ? ` =${widthParam}x${heightParam}` : '';
 
-    // while this isn't really a standard, it is common enough
-    if (yarleOptions.keepImageSize === OutputFormat.StandardMD || yarleOptions.keepImageSize === OutputFormat.LogSeqMD) {
+      // while this isn't really a standard, it is common enough
+      if (yarleOptions.imageSizeFormat === ImageSizeFormat.StandardMD) {
 
-      return `![](${realValue}${sizeString})`;
-    } else if (yarleOptions.keepImageSize === OutputFormat.ObsidianMD) {
-      sizeString = (widthParam || heightParam) ? `${widthParam || 0}x${heightParam || 0}` : '';
-      if (realValue.startsWith('./') || realValue.startsWith('..')) {
-        return sizeString != '' ? `![[${realValue}\\|${sizeString}]]` : `![[${realValue}${sizeString}]]`;
-      } else {
-        return `![${sizeString}](${realValue})`;
+        return `![](${realValue}${sizeString})`;
+      } else if (yarleOptions.imageSizeFormat === ImageSizeFormat.ObsidianMD) {
+        sizeString = (widthParam || heightParam) ? `${widthParam || 0}x${heightParam || 0}` : '';
+        if (realValue.startsWith('./') || realValue.startsWith('..')) {
+          return sizeString != '' ? `![[${realValue}\\|${sizeString}]]` : `![[${realValue}${sizeString}]]`;
+        } else {
+          return `![${sizeString}](${realValue})`;
+        }
       }
     }
-
     const useObsidianMD = isHeptaOrObsidianOutput();
     if (useObsidianMD && !value.match(/^[a-z]+:/)) {
       return `![[${realValue}]]`;