+++ test/yarle-special-cases.spec.ts
@@ -468,6 +468,121 @@ dateFormat: undefined,
 
   });
 
+  it('normalized notebook name', async () => {
+
+    const options: YarleOptions = {
+dateFormat: undefined,
+      enexSources: [ `${testDataFolder}test-#[notebook]^.enex` ],
+      templateFile: `${testDataFolder}keephtml-template.tmpl`,
+      outputDir: 'out',
+    };
+
+    await yarle.dropTheRope(options);
+    assert.equal(
+      fs.existsSync(
+        `${__dirname}/../out/notes/test-notebook/test -note with text only.md`,
+      ),
+      true,
+    );
+  
+  });
+
+  it('Multiple sanitized enex files with interEnex links', async () => {
+    const options: YarleOptions = {
+dateFormat: undefined,
+      enexSources: [ `${process.cwd()}${path.sep}test${path.sep}data${path.sep}SanitizedLinkedInterNotebooks` ],
+      outputDir: 'out',
+      isMetadataNeeded: true,
+      plainTextNotesOnly: false,
+      addExtensionToInternalLinks: true,
+      templateFile: undefined,
+      outputFormat: OutputFormat.StandardMD,
+
+    };
+    await dropTheRopeRunner.run(options);
+
+    // tslint:disable-next-line:no-console
+    console.log(`out dir: ${__dirname}/../out/notes/test-internotebook_links_A/Note2.md`);
+    assert.equal(
+      fs.existsSync(`${__dirname}/../out/notes/test-internotebook_links_A/Note2.md`),
+      true,
+    );
+
+    assert.equal(
+      eol.auto(fs.readFileSync(
+        `${__dirname}/../out/notes/test-internotebook_links_A/Note2.md`,
+        'utf8',
+      )),
+      fs.readFileSync(`${__dirname}/data/LinkedInterNotebooks/test-internotebook_links_A/Note2.md`, 'utf8'),
+    );
+
+    assert.equal(
+      fs.existsSync(`${__dirname}/../out/notes/test-internotebook_links_A/Table of Contents.md`),
+      true,
+    );
+
+    assert.equal(
+      eol.auto(fs.readFileSync(
+        `${__dirname}/../out/notes/test-internotebook_links_A/Table of Contents.md`,
+        'utf8',
+      )),
+      fs.readFileSync(`${__dirname}/data/LinkedInterNotebooks/test-internotebook_links_A/Table of Contents.md`, 'utf8'),
+    );
+
+    assert.equal(
+      fs.existsSync(`${__dirname}/../out/notes/test-internotebook_links_A/Note in Notebook A.md`),
+      true,
+    );
+
+    assert.equal(
+      eol.auto(fs.readFileSync(
+        `${__dirname}/../out/notes/test-internotebook_links_A/Note in Notebook A.md`,
+        'utf8',
+      )),
+      fs.readFileSync(`${__dirname}/data/LinkedInterNotebooks/test-internotebook_links_A/Note in Notebook A.md`, 'utf8'),
+    );
+
+    assert.equal(
+      fs.existsSync(`${__dirname}/../out/notes/test-internotebook_links_B/Untitled.md`),
+      true,
+    );
+
+    assert.equal(
+      eol.auto(fs.readFileSync(
+        `${__dirname}/../out/notes/test-internotebook_links_B/Untitled.md`,
+        'utf8',
+      )),
+      fs.readFileSync(`${__dirname}/data/LinkedInterNotebooks/test-internotebook_links_B/Untitled.md`, 'utf8'),
+    );
+
+    assert.equal(
+      fs.existsSync(`${__dirname}/../out/notes/test-internotebook_links_B/Table of Contents.md`),
+      true,
+    );
+
+    assert.equal(
+      eol.auto(fs.readFileSync(
+        `${__dirname}/../out/notes/test-internotebook_links_B/Table of Contents.md`,
+        'utf8',
+      )),
+      fs.readFileSync(`${__dirname}/data/LinkedInterNotebooks/test-internotebook_links_B/Table of Contents.md`, 'utf8'),
+    );
+
+    assert.equal(
+      fs.existsSync(`${__dirname}/../out/notes/test-internotebook_links_B/Note in Notebook B.md`),
+      true,
+    );
+
+    assert.equal(
+      eol.auto(fs.readFileSync(
+        `${__dirname}/../out/notes/test-internotebook_links_B/Note in Notebook B.md`,
+        'utf8',
+      )),
+      fs.readFileSync(`${__dirname}/data/LinkedInterNotebooks/test-internotebook_links_B/Note in Notebook B.md`, 'utf8'),
+    );
+
+  });
+
   it('Enex file with note containing text and picture', async () => {
     const options: YarleOptions = {
 dateFormat: undefined,
@@ -581,6 +696,57 @@ dateFormat: undefined,
         'utf8',
       )),
       fs.readFileSync(`${__dirname}/data/test-textWithImage.md`, 'utf8'),
+    );
+  });
+  it('Missing links', async () => {
+    const options: YarleOptions = {
+      dateFormat: undefined,
+      enexSources: [ `${testDataFolder}missing-link.enex` ],
+      outputDir: 'out',
+      isMetadataNeeded: true,
+      
+      keepEvernoteLinkIfNoNoteFound: true
+    };
+    await dropTheRopeRunner.run(options);
+    assert.equal(
+      fs.existsSync(
+        `${__dirname}/../out/notes/missing-link/NoteA.md`,
+      ),
+      true,
+    );
+
+
+    assert.equal(
+      eol.auto(fs.readFileSync(
+        `${__dirname}/../out/notes/missing-link/NoteA.md`,
+        'utf8',
+      )),
+      fs.readFileSync(`${__dirname}/data/test-missing-link.md`, 'utf8'),
+    );
+  });
+  it('Missing links - no keep ', async () => {
+    const options: YarleOptions = {
+      dateFormat: undefined,
+      enexSources: [ `${testDataFolder}missing-link-nokeep.enex` ],
+      outputDir: 'out',
+      isMetadataNeeded: true,
+      
+    };
+    await dropTheRopeRunner.run(options);
+    assert.equal(
+      fs.existsSync(
+        `${__dirname}/../out/notes/missing-link-nokeep/NoteA.md`,
+      ),
+      true,
+    );
+
+
+    assert.equal(
+      eol.auto(fs.readFileSync(
+        `${__dirname}/../out/notes/missing-link-nokeep/NoteA.md`,
+        'utf8',
+      )),
+      fs.readFileSync(`${__dirname}/data/test-missing-link-nokeep.md`, 'utf8'),
     );
   });
   it('Enex file with multiple notes', async () => {
@@ -973,7 +1139,7 @@ dateFormat: undefined,
 
   it('Multiple enex files with interEnex links', async () => {
     const options: YarleOptions = {
-dateFormat: undefined,
+      dateFormat: undefined,
       enexSources: [ `${process.cwd()}${path.sep}test${path.sep}data${path.sep}LinkedInterNotebooks` ],
       outputDir: 'out',
       isMetadataNeeded: true,
@@ -1066,6 +1232,94 @@ dateFormat: undefined,
     );
 
   });
+
+  it('Multiple untitled note - unregognized guid/link', async () => {
+    const options: YarleOptions = {
+      templateFile: `${testDataFolder}evernotelink-template.templ`,
+      enexSources: [ `${testDataFolder}two-untitled-notes-with-toc.enex` ],
+        "isMetadataNeeded": true,
+        "isNotebookNameNeeded": false,
+        "isZettelkastenNeeded": false,
+        "useZettelIdAsFilename": false,
+        "plainTextNotesOnly": false,
+        "skipLocation": true,
+        "skipCreationTime": true,
+        "skipUpdateTime": true,
+        "skipSourceUrl": true,
+        "skipWebClips": false,
+        "skipTags": true,
+        "useHashTags": true,
+        "outputFormat": OutputFormat.ObsidianMD,
+        "taskOutputFormat": TaskOutputFormat.ObsidianMD,
+        "skipEnexFileNameFromOutputPath": false,
+        "keepMDCharactersOfENNotes": false,
+        "monospaceIsCodeBlock": false,
+        "keepOriginalHtml": true,
+        "resourcesDir": "_resources",
+        "trimStartingTabs": false,
+        "convertPlainHtmlNewlines": false,
+        "encryptionPasswords": [
+          ""
+        ],
+        "logseqSettings": {
+          "journalNotes": false
+        },
+        "obsidianSettings": {
+          "omitLinkDisplayName": false
+        },
+        "dateFormat": "YYYY-MM-DD HH:mm",
+        "imageSizeFormat": OutputFormat.ObsidianMD,
+        "keepImageSize": true,
+        "keepOriginalAmountOfNewlines": true,
+        "addExtensionToInternalLinks": false,
+        "generateNakedUrls": true,
+        "urlEncodeFileNamesAndLinks": false,
+        "haveEnexLevelResources": false,
+        "haveGlobalResources": false,
+        "useUniqueUnknownFileNames": true,
+        "useLevenshteinForLinks": false,
+        "sanitizeResourceNameSpaces": false,
+        "replacementChar": "_",
+        "replacementCharacterMap": {
+          "<": "_",
+          ">": "_",
+          ":": "_",
+          "\"": "_",
+          "/": "_",
+          "\\": "_",
+          "|": "_",
+          "?": "_",
+          "*": "_"
+        }
+      
+    };
+    await dropTheRopeRunner.run(options);
+
+    assert.equal(
+      fs.existsSync(`${__dirname}/../out/notes/two-untitled-notes-with-toc/Untitled Note.1.md`),
+      true,
+    );
+    assert.equal(
+      fs.existsSync(`${__dirname}/../out/notes/two-untitled-notes-with-toc/Untitled Note.md`),
+      true,
+    );
+
+    assert.equal(
+      eol.auto(fs.readFileSync(
+        `${__dirname}/../out/notes/two-untitled-notes-with-toc/Untitled Note.md`,
+        'utf8',
+      )),
+      fs.readFileSync(`${__dirname}/data/two-untitled-notes-with-toc.md`, 'utf8'),
+    );
+    assert.equal(
+      eol.auto(fs.readFileSync(
+        `${__dirname}/../out/notes/two-untitled-notes-with-toc/Untitled Note.1.md`,
+        'utf8',
+      )),
+      fs.readFileSync(`${__dirname}/data/two-untitled-notes-with-toc.1.md`, 'utf8'),
+    );
+
+  });
   it('Enex file with PDF attachment', async () => {
     const options: YarleOptions = {
 dateFormat: undefined,
@@ -1147,7 +1401,8 @@ dateFormat: undefined,
         logseqSettings: { journalNotes: false },
         obsidianSettings: { omitLinkDisplayName: false },
         dateFormat: "YYYY-MM-DD hh:mm",
-        keepImageSize: OutputFormat.ObsidianMD,
+        imageSizeFormat: OutputFormat.ObsidianMD,
+        keepImageSize: true,
         keepOriginalAmountOfNewlines: true,
         addExtensionToInternalLinks: true,
         generateNakedUrls: false,
@@ -1209,7 +1464,8 @@ dateFormat: undefined,
         logseqSettings: { journalNotes: false },
         obsidianSettings: { omitLinkDisplayName: false },
         dateFormat: "YYYY-MM-DD hh:mm",
-        keepImageSize: OutputFormat.ObsidianMD,
+        imageSizeFormat: OutputFormat.ObsidianMD,
+        keepImageSize: true,
         keepOriginalAmountOfNewlines: true,
         addExtensionToInternalLinks: true,
         generateNakedUrls: false,
@@ -1273,7 +1529,8 @@ dateFormat: undefined,
         logseqSettings: { journalNotes: false },
         obsidianSettings: { omitLinkDisplayName: false },
         dateFormat: "YYYY-MM-DD hh:mm",
-        keepImageSize: OutputFormat.ObsidianMD,
+        imageSizeFormat: OutputFormat.ObsidianMD,
+        keepImageSize: true,
         keepOriginalAmountOfNewlines: true,
         addExtensionToInternalLinks: true,
         generateNakedUrls: false,
@@ -1311,7 +1568,8 @@ dateFormat: undefined,
       isMetadataNeeded: true,
       plainTextNotesOnly: false,
       urlEncodeFileNamesAndLinks: true,
-      keepImageSize: OutputFormat.ObsidianMD,
+      imageSizeFormat: OutputFormat.ObsidianMD,
+      keepImageSize: true,
       outputFormat: OutputFormat.ObsidianMD
     }